refactor: Set results on request rather than locals
This refactor moves setting results for a template to the request
object instead of on `res.locals`.

This is to start moving away from having `res.locals` set and used
in lots of middleware as then it's harder to trace back from the view
where things are set.

The idea is to set things within the render function so that it is
easier to see their origin.

This begins to address #451.
