+++ app/moves/controllers/list-as-cards.js
@@ -1,32 +1,18 @@
 const { get } = require('lodash')
 
 const permissions = require('../../../common/middleware/permissions')
-const presenters = require('../../../common/presenters')
 
 module.exports = function listAsCards(req, res) {
-  const {
-    cancelledMovesByDate = [],
-    activeMovesByDate = [],
-    fromLocationId,
-  } = res.locals
-  const userPermissions = get(req.session, 'user.permissions')
+  const { params, resultsAsCards, session } = req
+  const userPermissions = get(session, 'user.permissions')
   const canViewMove = permissions.check('move:view', userPermissions)
   const template =
-    canViewMove && fromLocationId
+    canViewMove && params.locationId
       ? 'moves/views/list-as-cards'
       : 'moves/views/download'
   const locals = {
+    resultsAsCards,
     pageTitle: 'moves::dashboard.outgoing_moves',
-    resultsAsCards: {
-      active: presenters.movesByToLocation(activeMovesByDate),
-      cancelled: cancelledMovesByDate.map(
-        presenters.moveToCardComponent({
-          showMeta: false,
-          showTags: false,
-          showImage: false,
-        })
-      ),
-    },
   }
 
   res.render(template, locals)
